requirements.txt
scrapy==2.11.0
pymongo==4.6.1
python-dotenv==1.0.0
